@kokes 2024-03-06
Oh never mind, I got the tests to run against a fresh kind cluster and they pass.