@monkeynator 2024-04-05
:tada: This PR is included in version 3.13.0 :tada:

The release is available on [GitHub release](https://github.com/enix/x509-certificate-exporter/releases/tag/untagged-2047105cea17249edfdd)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: